Add more matrix/quaternion utilities
Bump project version to 1.8.0-SNAPSHOT and set compile source/targetCompatibility to Java 17. Add IMatrixMath.trace() and implement trace() in BaseDoubleMatrix and BaseFloatMatrix. Introduce matrix decomposition helpers for 3x3 and 4x4 matrices (Matrix3{d,f}.decompose, Matrix4{d,f}.decompose) and reuse them to extract translation/rotation/scale, including quaternion extraction from rotation matrices. Expand IQuaternionMath with many mutable and default operations (in-place arithmetic, normalize/tryNormalize, length, vector/euler/transform helpers) and adapt Quaterniond/Quaternionf implementations accordingly (new constructors, in-place ops, normalization, euler/axis conversions, conj, transform, etc.). Misc: minor null-check message fixes and import adjustments. These changes add richer quaternion math and matrix decomposition utilities.

+++ UnifiedVectors/src/main/java/de/m_marvin/univec/impl/Vec4d.java
@@ -292,10 +292,10 @@ public class Vec4d implements IVector4Math<Double, Vec4d> {
 	
 	@Override
 	public Vec4d clampI(Double min, Double max) {
-		this.x = Math.max((Double) min, Math.min(this.x, (Double) max));
-		this.y = Math.max((Double) min, Math.min(this.y, (Double) max));
-		this.z = Math.max((Double) min, Math.min(this.z, (Double) max));
-		this.w = Math.max((Double) min, Math.min(this.w, (Double) max));
+		this.x = Math.max(min, Math.min(this.x, max));
+		this.y = Math.max(min, Math.min(this.y, max));
+		this.z = Math.max(min, Math.min(this.z, max));
+		this.w = Math.max(min, Math.min(this.w, max));
 		return this;
 	}
 
@@ -381,7 +381,7 @@ public class Vec4d implements IVector4Math<Double, Vec4d> {
 	
 	@Override
 	public Double length() {
-		return (Double) Math.sqrt(this.lengthSqr());
+		return Math.sqrt(this.lengthSqr());
 	}
 
 	@Override
